mintersdk/requirements.txt
two1
rlp
pycryptodome
bitcoin


#
# Successfull versions
#
# two1==3.10.9
# rlp==1.0.3
# pycryptodome==3.7.0
# bitcoin==1.1.42